+++ src/net/sourceforge/kolmafia/AreaCombatData.java
@@ -410,7 +410,7 @@ public class AreaCombatData {
   }
 
   public int getWeighting(final MonsterData monster) {
-    int raw = (this.currentWeightings.getOrDefault(monster, 0)).intValue();
+    int raw = this.currentWeightings.getOrDefault(monster, 0);
     if (((raw >> (KoLCharacter.getAscensions() & 1)) & 1) == 0) {
       return -2; // impossible this ascension
     }
@@ -1054,7 +1054,7 @@ public class AreaCombatData {
         }
       }
 
-      double stealRate = Math.min(pocketRates.get(i).doubleValue() * pocketModifier, 1.0);
+      double stealRate = Math.min(pocketRates.get(i) * pocketModifier, 1.0);
       int rawDropRate = item.getCount() >> 16;
       double dropRate = Math.min(rawDropRate * (itemModifier + itemBonus), 100.0);
       double effectiveDropRate = stealRate * 100.0 + (1.0 - stealRate) * dropRate;
