+++ src/net/sourceforge/kolmafia/MonsterData.java
@@ -500,7 +500,7 @@ public class MonsterData extends AdventureResult {
   private int evaluate(Object obj, int value) {
     if (obj != null) {
       if (obj instanceof Integer) {
-        return ((Integer) obj).intValue();
+        return (Integer) obj;
       }
       if (obj instanceof String) {
         obj = compile(obj);
@@ -567,11 +567,11 @@ public class MonsterData extends AdventureResult {
       return 0;
     }
     if (this.health instanceof Integer) {
-      int hp = ((Integer) this.health).intValue();
+      int hp = (Integer) this.health;
 
       if (hp == 0
           && (this.attack == null
-              || (this.attack instanceof Integer && ((Integer) this.attack).intValue() == 0))) {
+              || (this.attack instanceof Integer && (Integer) this.attack == 0))) {
         // The monster is unknown, so do not apply modifiers
         return 0;
       }
@@ -605,9 +605,7 @@ public class MonsterData extends AdventureResult {
   public int getBaseHP() {
     return this.scale != null
         ? -1
-        : this.health == null
-            ? 0
-            : this.health instanceof Integer ? ((Integer) (this.health)).intValue() : -1;
+        : this.health == null ? 0 : this.health instanceof Integer ? (Integer) (this.health) : -1;
   }
 
   public int getAttack() {
@@ -628,8 +626,8 @@ public class MonsterData extends AdventureResult {
       return 0;
     }
     if (this.attack instanceof Integer) {
-      int attack = ((Integer) this.attack).intValue();
-      if (attack == 0 && ((Integer) this.health).intValue() == 0) {
+      int attack = (Integer) this.attack;
+      if (attack == 0 && (Integer) this.health == 0) {
         // The monster is unknown, so do not apply modifiers
         return 0;
       }
@@ -665,9 +663,7 @@ public class MonsterData extends AdventureResult {
   public int getBaseAttack() {
     return this.scale != null
         ? -1
-        : this.attack == null
-            ? 0
-            : this.attack instanceof Integer ? ((Integer) (this.attack)).intValue() : -1;
+        : this.attack == null ? 0 : this.attack instanceof Integer ? (Integer) (this.attack) : -1;
   }
 
   public int getDefense() {
@@ -690,8 +686,8 @@ public class MonsterData extends AdventureResult {
       return 0;
     }
     if (this.defense instanceof Integer) {
-      int defense = ((Integer) this.defense).intValue();
-      if (defense == 0 && ((Integer) this.health).intValue() == 0) {
+      int defense = (Integer) this.defense;
+      if (defense == 0 && (Integer) this.health == 0) {
         // The monster is unknown, so do not apply modifiers
         return 0;
       }
@@ -733,7 +729,7 @@ public class MonsterData extends AdventureResult {
         ? -1
         : this.defense == null
             ? 0
-            : this.defense instanceof Integer ? ((Integer) (this.defense)).intValue() : -1;
+            : this.defense instanceof Integer ? (Integer) (this.defense) : -1;
   }
 
   public int getRawInitiative() {
@@ -743,7 +739,7 @@ public class MonsterData extends AdventureResult {
   public int getBaseInitiative() {
     return this.initiative == null
         ? 0
-        : this.initiative instanceof Integer ? ((Integer) (this.initiative)).intValue() : -1;
+        : this.initiative instanceof Integer ? (Integer) (this.initiative) : -1;
   }
 
   public int getInitiative() {
